ai-service/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LangChain
langchain==0.1.0
langchain-anthropic==0.1.0

# HTTP clients
httpx==0.25.2
aiohttp==3.9.1

# Weaviate client
weaviate-client==3.25.3

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
tenacity==8.2.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
